flask_google_pre_trained/requirements.txt
gunicorn
Flask
nltk
request
google-cloud
google-cloud-language
jsonify